refactoring HomeViewModel
