+++ DesenheAi/View/ContentView.swift
@@ -31,7 +31,7 @@ struct ContentView: View {
     @State private var showingScore = false
     @State private var showingMessagePhoto = false
     
-    @State private var canvasView = PKCanvasView()
+    @State var canvasView = PKCanvasView()
     @State private var image: UIImage = UIImage()
     @State var previewDrawing: PKDrawing? = nil
     
@@ -43,6 +43,8 @@ struct ContentView: View {
     @State private var guideResultBR: String?
     
     var body: some View {
+        let homeViewModel = HomeViewModel(currentImage: $currentImage, image: $image, showingUS: $showingUS, guideResultUS: $guideResultUS, score: $score, scoreTitle: $scoreTitle, showingScore: $showingScore, isActive: $isActive, counter: $counter, canvasView: $canvasView)
+        
         NavigationView {
             ZStack {
                 Color(red: 219/255, green: 221/255, blue: 228/255, opacity: 1)
@@ -50,15 +52,15 @@ struct ContentView: View {
                 VStack(spacing: 15.0) {
                     Spacer(minLength: 20)
                     HStack (spacing: 10) {
-                        GrayCircleButton(actionButton: onClearTapped(), imagetext: "trash")
+                        GrayCircleButton(function: {homeViewModel.onClearTapped()}, imageText: "trash")
                             .padding(.leading, 20)
-                        GrayCircleButton(actionButton: imageGalery(), imagetext: "download-image")
+                        GrayCircleButton(function: {homeViewModel.imageGalery()}, imageText: "download-image")
                         Spacer()
                         TimerView(counter: $counter)
                     }
                     VStack {
                         ChallengeLabel(showingUS: $showingUS, guideResultUS: $guideResultUS, currentImage: $currentImage, isTop: true)
-                        CanvasView(canvasView: $canvasView, onSaved: onSaved)
+                        CanvasView(canvasView: $canvasView, onSaved: homeViewModel.onSaved)
                             .frame(width: UIScreen.main.bounds.maxX, height: UIScreen.main.bounds.maxX, alignment: .center)
                             .cornerRadius(20.0)
                             .background(Color.clear)
@@ -68,7 +70,7 @@ struct ContentView: View {
                     VStack {
                         Spacer()
                         Button(showingUS ? "New Challenge" : "Próximo desafio") {
-                            newChallenge()
+                            homeViewModel.newChallenge()
                         }
                         .titleNextButton()
                         .padding(.top, 55)
@@ -78,7 +80,7 @@ struct ContentView: View {
             .onAppear {
                 self.score = Int16(items.endIndex)
                 isActive = true
-                newChallenge()
+                homeViewModel.newChallenge()
             }
             .onReceive(timer) { time in
                 guard self.isActive else { return }
@@ -89,7 +91,7 @@ struct ContentView: View {
             .alert(isPresented: $showingScore) {
                 Alert(title: Text(scoreTitle), message: Text(showingUS ? "Seu escore é : \(score)" :  "Your score is \(score)"), dismissButton: .default(Text("Continue")) {
                     addItem()
-                    newChallenge()
+                    homeViewModel.newChallenge()
                 })
             }
             .onReceive(NotificationCenter.default.publisher(for: UIApplication.willResignActiveNotification)) { _ in
@@ -130,91 +132,6 @@ struct ContentView: View {
     }
 
 }
-
-// MARK: - Game methods
-
-private extension ContentView {
-    func onClearTapped() {
-        canvasView.drawing = PKDrawing()
-    }
-
-    func onSaved() {
-        image = canvasView.drawing.image(from: canvasView.bounds, scale: UIScreen.main.scale)
-        checkChallenge()
-    }
-    
-    func imageGalery() {
-        UIImageWriteToSavedPhotosAlbum(image, nil, nil, nil)
-    }
-    
-    func newChallenge(){
-        counter = 0
-        isActive = true
-        let index = Int.random(in: 0...200)
-        onClearTapped()
-        guideResultUS = LabelUSView.getChallenge(index)
-        showingScore = false
-    }
-    
-    func checkAnswer() {
-        if currentImage == guideResultUS {
-            score += 1
-            if showingUS {
-                scoreTitle = "🏆 Great!"
-            } else {
-                scoreTitle = "🏆 Correto!"
-          }
-        showingScore = true
-        } else {
-            if TimerViewModel.completed(counter) {
-                if showingUS {
-                    scoreTitle = "⌛️ Wrong. Time is over. Start again..."
-                } else {
-                    scoreTitle = "⌛️ O tempo acabou. Vamos mais uma vez?"
-                }
-                showingScore = true
-                isActive = false
-            }
-        }
-    }
-    
-    func checkChallenge() {
-        let config = MLModelConfiguration()
-        guard let coreMLModel = try? Desenheai(configuration: config),
-              let visionModel = try? VNCoreMLModel(for: coreMLModel.model) else {return}
-        
-        let request = VNCoreMLRequest(model: visionModel, completionHandler: { [] request, error in
-            if let sortedResults = request.results! as? [VNClassificationObservation] {
-                let topResult = sortedResults.first
-                   DispatchQueue.main.async {
-                    for result in sortedResults {
-                        if result.confidence < 0.1 {
-                                self.currentImage = "something strange"
-                        }
-                        self.currentImage = "\(topResult?.identifier ?? "algo estranho")"
-                        self.checkAnswer()
-                        print(result.identifier, result.confidence)
-                        print("-----------------------")
-                         }
-                              
-                   }
-         } })
-        request.imageCropAndScaleOption = .centerCrop
-        
-        DispatchQueue.global(qos: .userInitiated).async {
-            let grayscaleImage = image
-            
-            let handler = VNImageRequestHandler(cgImage: grayscaleImage.cgImage!, options: [:])
-            do {
-                try handler.perform([request])
-                
-            } catch {
-                print("Failed to perform classification.\n\(error.localizedDescription)")
-            }
-        }
-    }
-}
-
 struct ContentView_Previews: PreviewProvider {
     static var previews: some View {
         ContentView().environment(\.managedObjectContext, PersistenceController.preview.container.viewContext)
